% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{evidence}
\alias{evidence}
\title{Get supported evidence types}
\source{
https://wiki.thebiogrid.org/doku.php/biogridrest
}
\usage{
evidence(accessKey = NULL)
}
\arguments{
\item{accessKey}{A \code{character}. Access Keys are free and openly
available at \url{https://webservice.thebiogrid.org/}. Defaulit is NULL.}
}
\value{
A \code{tibble}
\describe{
\item{evidenceList}{Supported evidence type.}
}
}
\description{
Get supported evidence types
}
