% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{identifiers}
\alias{identifiers}
\title{Get supported identifiers}
\source{
https://wiki.thebiogrid.org/doku.php/biogridrest
}
\usage{
identifiers(accessKey = NULL)
}
\arguments{
\item{accessKey}{A \code{character}. Access Keys are free and openly
available at \url{https://webservice.thebiogrid.org/}. Defaulit is NULL.}
}
\value{
A \code{tibble}
\describe{
\item{additionalIdentifierTypes}{Supported identifier type.}
}
}
\description{
Get supported identifiers
}
