% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{interactions}
\alias{interactions}
\title{Get Interactions}
\source{
https://wiki.thebiogrid.org/doku.php/biogridrest

https://wiki.thebiogrid.org/doku.php/downloads
}
\usage{
interactions(accessKey = NULL, geneList = NULL, searchNames = TRUE,
  taxId, start = 0, max = 5, htpThreshold = 0,
  interSpeciesExcluded = FALSE, selfInteractionsExcluded = FALSE,
  includeEvidence = FALSE, searchIds = FALSE, searchSynonyms = FALSE,
  searchBiogridIds = FALSE, excludeGenes = FALSE,
  includeInteractors = FALSE, includeInteractorInteractions = FALSE,
  excludePubmeds = FALSE, throughputTag = "any", format = "tab2",
  includeHeader = TRUE, pubmedList, evidenceList,
  additionalIdentifierTypes)
}
\arguments{
\item{accessKey}{A \code{character}. Access Keys are free and openly
available at \url{https://webservice.thebiogrid.org/}. Defaulit is NULL.}

\item{geneList}{A \code{character} vector. Interactions between genes in
this list will be fetched. Default is NULL.}

\item{searchNames}{A \code{logical}. If \code{TRUE} (default), the
interactor OFFICIAL_SYMBOL will be examined for a match with the
\code{geneList}.}

\item{taxId}{A \code{numeric}. NCBI taxonomy identifiers.
The full list of supported organisms can be obtained by calling
\code{\link{organisms}}.}

\item{start}{A \code{numeric}. Results fetched will start at this value
d(efault 0).}

\item{max}{A \code{numeric}. Results fetched will end at this value range
from 0 to 10000 (default is 5).}

\item{htpThreshold}{A \code{logical}. Interactions whose Pubmed ID has more
than this number of interactions will be excluded from the results (default
is \code{FALSE}). Ignored if \code{excludePubmeds} is \code{FALSE}.}

\item{interSpeciesExcluded}{A \code{logical}. If \code{TRUE}, interactions
with interactors from different species will be excluded (default
\code{FALSE}).}

\item{selfInteractionsExcluded}{A \code{logical}. If \code{TRUE},
interactions with one interactor will be excluded (default \code{FALSE}).}

\item{includeEvidence}{A \code{logical}. If \code{TRUE}, any interaction
evidence with its Experimental System in the \code{evidenceList} will be
included in the result (default \code{FALSE}).}

\item{searchIds}{A \code{logical}. If \code{TRUE}, the interactor
ENTREZ_GENE, ORDERED LOCUS and SYSTEMATIC_NAME (orf) will be examined for a
match with the \code{geneList} (default \code{FALSE}).}

\item{searchSynonyms}{A \code{logical}. If \code{TRUE}, the interactor
SYNONYM will be examined for a match with the \code{geneList} (default
\code{FALSE}).}

\item{searchBiogridIds}{A \code{logical}. If \code{TRUE}, the entries in
'GENELIST' will be compared to BIOGRID internal IDS which are provided in
 all Tab2 formatted files (default \code{FALSE}).}

\item{excludeGenes}{A \code{logical}. If \code{TRUE}, interactions
containing genes in the \code{geneList} will be excluded from the results.
Ignored if one of \code{searchIds}, \code{searchNames},
\code{searchSynonyms} is not \code{TRUE} and
\code{additionalIdentifierTypes} is empty (default \code{FALSE}).}

\item{includeInteractors}{A \code{logical}. If \code{TRUE}, in addition to
interactions between genes on the \code{geneList}, interactions will also
be fetched which have only one interactor on the \code{geneList} (default
\code{FALSE}).}

\item{includeInteractorInteractions}{A \code{logical}. If \code{TRUE},
interactions between the \code{geneList}’s first order interactors will be
included (default \code{FALSE}). Ignored if \code{includeInteractors} is
\code{FALSE} or if \code{excludeGenes} is set to \code{TRUE}.}

\item{excludePubmeds}{A \code{logical}. If \code{FALSE} (default),
interactions with Pubmed ID in \code{pubmedList} will be included in the
results.}

\item{throughputTag}{A \code{character}. If set to 'low' or 'high', only
interactions with 'Low throughput' or 'High throughput' in the 'throughput'
field will be returned (default is 'any').}

\item{format}{A \code{character} only 'tab2' is used.}

\item{includeHeader}{A \code{logical}. If \code{TRUE}, the first line of the
result will be a BioGRID column header (always \code{TRUE}).}

\item{pubmedList}{A \code{character} vector. Interactions will be fetched
whose Pubmed Id is/ is not in this list, depending on the value of
\code{excludePubmeds}.}

\item{evidenceList}{A \code{character} vector. Any interaction evidence with
its Experimental System in the list will be excluded from the results unless
\code{includeEvidence} is set to \code{TRUE}. The full list can be obtained
by calling \code{\link{evidence}}.}

\item{additionalIdentifierTypes}{A \code{character} vector. Identifier types
on this list are examined for a match with the \code{geneList}. The full
list can be obtained by calling \code{\link{identifiers}}.}
}
\value{
A \code{tibble}
\describe{
\item{BioGRID Interaction ID}{A unique identifier for each interaction within the BioGRID database. Can be used to link to BioGRID interaction pages. For example: http://thebiogrid.org/interaction/616539}
\item{Entrez Gene ID for Interactor A}{The identifier from the Entrez-Gene database that corresponds to Interactor A. If no Entrez Gene ID is available, this will be a “-”.}
\item{Entrez Gene ID for Interactor B}{Same structure as column 2.}
\item{BioGRID ID for Interactor A}{The identifier in the BioGRID database that corresponds to Interactor A. These identifiers are best used for creating links to the BioGRID from your own websites or applications. To link to a page within our site, simply append the URL: http://thebiogrid.org/ID/ to each ID. For example, http://thebiogrid.org/31623/.}
\item{BioGRID ID for Interactor B}{Same structure as column 4.}
\item{Systematic name for Interactor A}{A plain text systematic name if known for interactor A. Will be a “-” if no name is available.}
\item{Systematic name for Interactor B}{Same structure as column 6.}
\item{Official symbol for Interactor A}{A common gene name/official symbol for interactor A. Will be a “-” if no name is available.}
\item{Official symbol for Interactor B}{Same structure as column 8.}
\item{Synonyms/Aliases for Interactor A}{A “|” separated list of alternate identifiers for interactor A. Will be “-” if no aliases are available.}
\item{Synonyms/Aliases for Interactor B}{Same stucture as column 10.}
\item{Experimental System Name}{One of the many Experimental Evidence Codes supported by the BioGRID.}
\item{Experimental System Type}{This will be either “physical” or “genetic” as a classification of the Experimental System Name.}
\item{First author surname of the publication in which the interaction has been shown, optionally followed by additional indicators, e.g}{Stephenson A (2005)}
\item{Pubmed ID}{ of the publication in which the interaction has been shown.}
\item{Organism ID for Interactor A}{This is the NCBI Taxonomy ID for Interactor A.}
\item{Organism ID for Interactor B}{Same structure as 16.}
\item{Interaction Throughput}{This will be either High Throughput, Low Throughput or Both (separated by “|”).}
\item{Quantitative Score}{This will be a positive for negative value recorded by the original publication depicting P-Values, Confidence Score, SGA Score, etc. Will be “-” if no score is reported.}
\item{Post Translational Modification}{For any Biochemical Activity experiments, this field will be filled with the associated post translational modification. Will be “-” if no modification is reported.}
\item{Phenotypes}{If any phenotype info is recorded, it will be provided here separated by “|”. Each phenotype will be of the format <phenotype>[<phenotype qualifier>]:<phenotype type>. Note that the phenotype types and qualifiers are optional and will only be present where recorded. Phenotypes may also have multiple qualifiers in which case unique qualifiers will be separated by carat (^). If no phenotype information is available, this field will contain “-”.}
\item{Qualifications}{If additional plain text information was recorded for an interaction, it will be listed with unique qualifiers separated by “|”. If no qualification is available, this field will contain “-”.}
\item{Tags}{If an interaction has been tagged with additional classifications, they will be provided in this column separated by “|”. If no tag information is available, this field will contain “-”.}
\item{Source Database}{This field will contain the name of the database in which this interaction was provided.}
}
}
\description{
Get Interactions
}
\seealso{
identifiers

organisms

evidence
}
