% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{organisms}
\alias{organisms}
\title{Get supported organisms}
\source{
https://wiki.thebiogrid.org/doku.php/biogridrest
}
\usage{
organisms(accessKey = NULL)
}
\arguments{
\item{accessKey}{A \code{character}. Access Keys are free and openly
available at \url{https://webservice.thebiogrid.org/}. Defaulit is NULL.}
}
\value{
A \code{tibble}
\describe{
\item{taxId}{NCBI organism taxon ID.}
\item{taxName}{Organism name.}
}
}
\description{
Get supported organisms
}
